lint: force using browser pollyfill instead of chrome
organize imports

- changed "chrome" to "browser"  in background-events.js

- added "tabs" permission to firefox manifest
- moved event handler of URL_CHANGED to `content-events.js`
- added a new subscription method for url change in `content.js`

- added a test for issue #107

- moved EVENT_URL_CHANGED's event handler to the already existing listener.
- changed page locator in test.
- used a MutationObserver to insert `overlay-package-report` after the repository element is rendered (there are cases that the repository selector cannot be found since it's not yet rendered).

- fixed test

lint e2e too

- removed "highlight" and fixed test

- removed `tabs` permission from manifest
- added utility function that waits for element to be rendered
- rewrote test

ci: add discord notification (#108)

Added a discord notification workflow.
The file added was used for testing purposes, and if anyone else would
want to test it.
It may also be used to split up the notifying workflow from the other
build and publish workflow, and I guess more platform notifications will
be added overtime, but it is up to you.

The notification content is:
![Screenshot 2023-06-12 at 16 52
05](https://github.com/os-scar/overlay/assets/62609377/1c2b4e94-232a-4eab-8ecb-6e8e49b54fab)

**when creating the Discord webhook you can choose the name of the bot

---------

Co-authored-by: Baruch Odem (Rothkoff) <baruchiro@gmail.com>

fix real examples

lint e2e too

- removed "highlight" and fixed test

ci: add discord notification (#108)

Added a discord notification workflow.
The file added was used for testing purposes, and if anyone else would
want to test it.
It may also be used to split up the notifying workflow from the other
build and publish workflow, and I guess more platform notifications will
be added overtime, but it is up to you.

The notification content is:
![Screenshot 2023-06-12 at 16 52
05](https://github.com/os-scar/overlay/assets/62609377/1c2b4e94-232a-4eab-8ecb-6e8e49b54fab)

**when creating the Discord webhook you can choose the name of the bot

---------

Co-authored-by: Baruch Odem (Rothkoff) <baruchiro@gmail.com>

fix real examples

- fixed typo